// src/app/shared/models/sudoku.enum.ts
export enum SUDOKU {
  GAMES = 'games',
  PLAYERS = 'players',
  PLAYER_COUNT = 'playerCount',
  PUZZLE = '_sudokuPuzzle',
  SOLVED = '_sudokuSolved',
  ORIGINAL_PUZZLE = '_originalSudokuPuzzle',
  GAME_START_TIME = 'gameStartTime',
  GAME_END_TIME = 'gameEndTime',
  IS_TIMER_INITIALIZED = 'isTimerInitialized',
  DIFFICULTY = 'difficulty',
  IS_GAME_ENDED = 'isGameEnded',
}